chat: add icon badges, "inline" messages, and multiple classes on custom message types
If you pass an icon class name (eg. "icon-meh") as the `badge` property of a message
object, that icon will be shown in the badge box.
Inline messages are smaller than the normal messages, with the message body starting
on the same line as the user name. Also, the badge box is smaller and has a transparent
background.