chat-type: fix compatibility with other extensions
Other extensions would add an event handler to chat:receive to do
work _after_ the ChatView rendered the message, but ExtPlug
reattached the ChatView handler at the end of the handler list.
If ExtPlug was loaded after RCS or p3, this would break chat
entirely. This fixes it by replacing the handler in-place, instead
of reattaching it.